qcacld-3.0: STC misc fixes
Multiple miscellaneous fixes for STC feature.
1) Increment the packet count per RX flow in FISA.
2) Incorporate the updated Burst definition
[5000 bytes in 10ms to start the burst,
0 bytes for 1500ms to end the burst]
3) Modify the flow settling time threshold to 3 seconds
4) Fix the packet length calculations for TX and RX path,
to exclude the ethernet header.
5) Fix burst detection and state machine
6) Change the pkt_iat and burst_duration data-type to uint64_t
to account for time stored in nano-seconds.
7) Compilation error fix from I2e1b0d8d39f60e6c9dbf0bfcbd2e82a470e9ddd6

CRs-Fixed: 3817653

+++ components/dp/dispatcher/inc/wlan_dp_public_struct.h
@@ -662,8 +662,8 @@ struct wlan_dp_stc_flow_classify_result {
 struct wlan_dp_stc_txrx_min_max_stats {
 	uint32_t pkt_size_min;
 	uint32_t pkt_size_max;
-	uint32_t pkt_iat_min;
-	uint32_t pkt_iat_max;
+	uint64_t pkt_iat_min;
+	uint64_t pkt_iat_max;
 };
 
 /*
@@ -681,9 +681,9 @@ struct wlan_dp_stc_txrx_stats {
 	uint32_t pkts;
 	uint32_t pkt_size_min;
 	uint32_t pkt_size_max;
-	uint32_t pkt_iat_min;
-	uint32_t pkt_iat_max;
-	uint32_t pkt_iat_sum;
+	uint64_t pkt_iat_min;
+	uint64_t pkt_iat_max;
+	uint64_t pkt_iat_sum;
 };
 
 /*
@@ -709,9 +709,9 @@ struct wlan_dp_stc_txrx_samples {
  * @burst_count: Total number of bursts
  */
 struct wlan_dp_stc_burst_stats {
-	uint32_t burst_duration_min;
-	uint32_t burst_duration_max;
-	uint32_t burst_duration_sum;
+	uint64_t burst_duration_min;
+	uint64_t burst_duration_max;
+	uint64_t burst_duration_sum;
 	uint32_t burst_size_min;
 	uint32_t burst_size_max;
 	uint64_t burst_size_sum;
